bi-system/requirements.txt
Django>=4.2
djangorestframework
mssql-django
pyodbc
celery
redis
django-cors-headers
pandas
sqlalchemy
openpyxl
pymysql
psycopg2-binary
oracledb
